updated tests and fixed problem with casting negative values to unsigned types

+++ tests/linear_kernels/subtraction/subtraction_dcmpl.cl
@@ -285,33 +285,33 @@ void subtract_get_work_dim_get_group_id(int x, __global uint *data)
 __kernel __attribute__((reqd_work_group_size(16, 2, 2)))
 void subtract_x_get_local_size(__global int *data, int x)
 {
-    data[get_global_id(0)] = (ulong)x + (-16);
-    data[get_global_id(1)] = (ulong)x + (-2);
-    data[get_global_id(2)] = (ulong)x + (-2);
+    data[get_global_id(0)] = (ulong)x + (ulong)(-16);
+    data[get_global_id(1)] = (ulong)x + (ulong)(-2);
+    data[get_global_id(2)] = (ulong)x + (ulong)(-2);
 }
 
 __kernel __attribute__((reqd_work_group_size(2, 16, 2)))
 void subtract_get_global_offset_get_local_size(int x, __global int *data)
 {
-    data[get_global_id(0)] = (ulong)get_global_offset(0) + (-2);
-    data[get_global_id(1)] = (ulong)get_global_offset(1) + (-16);
-    data[get_global_id(2)] = (ulong)get_global_offset(2) + (-2);
+    data[get_global_id(0)] = (ulong)get_global_offset(0) + (ulong)(-2);
+    data[get_global_id(1)] = (ulong)get_global_offset(1) + (ulong)(-16);
+    data[get_global_id(2)] = (ulong)get_global_offset(2) + (ulong)(-2);
 }
 
 __kernel __attribute__((reqd_work_group_size(8, 4, 2)))
 void subtract_get_local_id_get_local_size(int x, __global int *data)
 {
-    data[get_global_id(0)] = (-8) + (ulong)get_local_id(0);
-    data[get_global_id(1)] = (-4) + (ulong)get_local_id(1);
-    data[get_global_id(2)] = (-2) + (ulong)get_local_id(2);
+    data[get_global_id(0)] = (ulong)(-8) + (ulong)get_local_id(0);
+    data[get_global_id(1)] = (ulong)(-4) + (ulong)get_local_id(1);
+    data[get_global_id(2)] = (ulong)(-2) + (ulong)get_local_id(2);
 }
 
 __kernel __attribute__((reqd_work_group_size(2, 16, 2)))
 void subtract_get_group_id_get_local_size(int x, __global int *data)
 {
-    data[get_global_id(0)] = (ulong)get_group_id(0) + (-2);
-    data[get_global_id(1)] = (ulong)get_group_id(1) + (-16);
-    data[get_global_id(2)] = (ulong)get_group_id(2) + (-2);
+    data[get_global_id(0)] = (ulong)get_group_id(0) + (ulong)(-2);
+    data[get_global_id(1)] = (ulong)get_group_id(1) + (ulong)(-16);
+    data[get_global_id(2)] = (ulong)get_group_id(2) + (ulong)(-2);
 }
 
 __kernel __attribute__((reqd_work_group_size(2, 2, 4)))
@@ -325,33 +325,33 @@ void subtract_get_local_size_get_local_size(int x, __global uint *data)
 __kernel __attribute__((reqd_work_group_size(2, 2, 4)))
 void subtract_get_global_id_get_local_size(int x, __global uint *data)
 {
-    data[get_global_id(0) - get_global_offset(0)] = (-2) + get_global_id(0);
-    data[get_global_id(1)] = (-2) + get_global_id(1);
-    data[get_global_id(2)] = (-4) + get_global_id(2);
+    data[get_global_id(0) - get_global_offset(0)] = (ulong)(-2) + get_global_id(0);
+    data[get_global_id(1)] = (ulong)(-2) + get_global_id(1);
+    data[get_global_id(2)] = (ulong)(-4) + get_global_id(2);
 }
 
 __kernel __attribute__((reqd_work_group_size(16, 2, 2)))
 void subtract_get_global_size_get_local_size(__global int *data, int x)
 {
-    data[get_global_id(0)] = (ulong)get_global_size(0) + (-16);
-    data[get_global_id(1)] = (ulong)get_global_size(1) + (-2);
-    data[get_global_id(2)] = (ulong)get_global_size(2) + (-2);
+    data[get_global_id(0)] = (ulong)get_global_size(0) + (ulong)(-16);
+    data[get_global_id(1)] = (ulong)get_global_size(1) + (ulong)(-2);
+    data[get_global_id(2)] = (ulong)get_global_size(2) + (ulong)(-2);
 }
 
 __kernel __attribute__((reqd_work_group_size(2, 2, 4)))
 void subtract_get_num_groups_get_local_size(int x, __global uint *data)
 {
-    data[get_global_id(0)] = (ulong)get_num_groups(0) + (-2);
-    data[get_global_id(1)] = (ulong)get_num_groups(1) + (-2);
-    data[get_global_id(2)] = (ulong)get_num_groups(2) + (-4);
+    data[get_global_id(0)] = (ulong)get_num_groups(0) + (ulong)(-2);
+    data[get_global_id(1)] = (ulong)get_num_groups(1) + (ulong)(-2);
+    data[get_global_id(2)] = (ulong)get_num_groups(2) + (ulong)(-4);
 }
 
 __kernel __attribute__((reqd_work_group_size(2, 2, 4)))
 void subtract_get_work_dim_get_local_size(int x, __global uint *data)
 {
-    data[get_global_id(0)] = get_work_dim() + (-2);
-    data[get_global_id(1)] = get_work_dim() + (-2);
-    data[get_global_id(2)] = get_work_dim() + (-4);
+    data[get_global_id(0)] = get_work_dim() + (ulong)(-2);
+    data[get_global_id(1)] = get_work_dim() + (ulong)(-2);
+    data[get_global_id(2)] = get_work_dim() + (ulong)(-4);
 }
 
 __kernel __attribute__((reqd_work_group_size(16, 2, 2)))
@@ -1037,33 +1037,33 @@ void subtract_long_get_work_dim_get_group_id(int x, __global ulong *data)
 __kernel __attribute__((reqd_work_group_size(16, 2, 2)))
 void subtract_long_x_get_local_size(__global long *data, int x)
 {
-    data[get_global_id(0)] = (ulong)x + (-16);
-    data[get_global_id(1)] = (ulong)x + (-2);
-    data[get_global_id(2)] = (ulong)x + (-2);
+    data[get_global_id(0)] = (ulong)x + (ulong)(-16);
+    data[get_global_id(1)] = (ulong)x + (ulong)(-2);
+    data[get_global_id(2)] = (ulong)x + (ulong)(-2);
 }
 
 __kernel __attribute__((reqd_work_group_size(2, 16, 2)))
 void subtract_long_get_global_offset_get_local_size(int x, __global long *data)
 {
-    data[get_global_id(0)] = (ulong)get_global_offset(0) + (-2);
-    data[get_global_id(1)] = (ulong)get_global_offset(1) + (-16);
-    data[get_global_id(2)] = (ulong)get_global_offset(2) + (-2);
+    data[get_global_id(0)] = (ulong)get_global_offset(0) + (ulong)(-2);
+    data[get_global_id(1)] = (ulong)get_global_offset(1) + (ulong)(-16);
+    data[get_global_id(2)] = (ulong)get_global_offset(2) + (ulong)(-2);
 }
 
 __kernel __attribute__((reqd_work_group_size(8, 4, 2)))
 void subtract_long_get_local_id_get_local_size(int x, __global long *data)
 {
-    data[get_global_id(0)] = (ulong)get_local_id(0) + (-8);
-    data[get_global_id(1)] = (ulong)get_local_id(1) + (-4);
-    data[get_global_id(2)] = (ulong)get_local_id(2) + (-2);
+    data[get_global_id(0)] = (ulong)get_local_id(0) + (ulong)(-8);
+    data[get_global_id(1)] = (ulong)get_local_id(1) + (ulong)(-4);
+    data[get_global_id(2)] = (ulong)get_local_id(2) + (ulong)(-2);
 }
 
 __kernel __attribute__((reqd_work_group_size(2, 16, 2)))
 void subtract_long_get_group_id_get_local_size(int x, __global long *data)
 {
-    data[get_global_id(0)] = (ulong)get_group_id(0) + (-2);
-    data[get_global_id(1)] = (ulong)get_group_id(1) + (-16);
-    data[get_global_id(2)] = (ulong)get_group_id(2) + (-2);
+    data[get_global_id(0)] = (ulong)get_group_id(0) + (ulong)(-2);
+    data[get_global_id(1)] = (ulong)get_group_id(1) + (ulong)(-16);
+    data[get_global_id(2)] = (ulong)get_group_id(2) + (ulong)(-2);
 }
 
 __kernel __attribute__((reqd_work_group_size(2, 2, 4)))
@@ -1077,33 +1077,33 @@ void subtract_long_get_local_size_get_local_size(int x, __global ulong *data)
 __kernel __attribute__((reqd_work_group_size(2, 2, 4)))
 void subtract_long_get_global_id_get_local_size(int x, __global ulong *data)
 {
-    data[get_global_id(0) - get_global_offset(0)] = get_global_id(0) + (-2);
-    data[get_global_id(1)] = get_global_id(1) + (-2);
-    data[get_global_id(2)] = get_global_id(2) + (-4);
+    data[get_global_id(0) - get_global_offset(0)] = get_global_id(0) + (ulong)(-2);
+    data[get_global_id(1)] = get_global_id(1) + (ulong)(-2);
+    data[get_global_id(2)] = get_global_id(2) + (ulong)(-4);
 }
 
 __kernel __attribute__((reqd_work_group_size(16, 2, 2)))
 void subtract_long_get_global_size_get_local_size(__global long *data, int x)
 {
-    data[get_global_id(0)] = (ulong)get_global_size(0) + (-16);
-    data[get_global_id(1)] = (ulong)get_global_size(1) + (-2);
-    data[get_global_id(2)] = (ulong)get_global_size(2) + (-2);
+    data[get_global_id(0)] = (ulong)get_global_size(0) + (ulong)(-16);
+    data[get_global_id(1)] = (ulong)get_global_size(1) + (ulong)(-2);
+    data[get_global_id(2)] = (ulong)get_global_size(2) + (ulong)(-2);
 }
 
 __kernel __attribute__((reqd_work_group_size(2, 2, 4)))
 void subtract_long_get_num_groups_get_local_size(int x, __global ulong *data)
 {
-    data[get_global_id(0)] = (ulong)get_num_groups(0) + (-2);
-    data[get_global_id(1)] = (ulong)get_num_groups(1) + (-2);
-    data[get_global_id(2)] = (ulong)get_num_groups(2) + (-4);
+    data[get_global_id(0)] = (ulong)get_num_groups(0) + (ulong)(-2);
+    data[get_global_id(1)] = (ulong)get_num_groups(1) + (ulong)(-2);
+    data[get_global_id(2)] = (ulong)get_num_groups(2) + (ulong)(-4);
 }
 
 __kernel __attribute__((reqd_work_group_size(2, 2, 4)))
 void subtract_long_get_work_dim_get_local_size(int x, __global ulong *data)
 {
-    data[get_global_id(0)] = get_work_dim() + (-2);
-    data[get_global_id(1)] = get_work_dim() + (-2);
-    data[get_global_id(2)] = get_work_dim() + (-4);
+    data[get_global_id(0)] = get_work_dim() + (ulong)(-2);
+    data[get_global_id(1)] = get_work_dim() + (ulong)(-2);
+    data[get_global_id(2)] = get_work_dim() + (ulong)(-4);
 }
 
 __kernel __attribute__((reqd_work_group_size(16, 2, 2)))
@@ -1805,17 +1805,17 @@ void subtract_char_get_global_offset_get_local_size(int x, __global char *data)
 __kernel __attribute__((reqd_work_group_size(8, 4, 2)))
 void subtract_char_get_local_id_get_local_size(int x, __global char *data)
 {
-    data[get_global_id(0)] = (-8) + (ulong)get_local_id(0);
-    data[get_global_id(1)] = (-4) + (ulong)get_local_id(1);
-    data[get_global_id(2)] = (-2) + (ulong)get_local_id(2);
+    data[get_global_id(0)] = (ulong)(-8) + (ulong)get_local_id(0);
+    data[get_global_id(1)] = (ulong)(-4) + (ulong)get_local_id(1);
+    data[get_global_id(2)] = (ulong)(-2) + (ulong)get_local_id(2);
 }
 
 __kernel __attribute__((reqd_work_group_size(2, 16, 2)))
 void subtract_char_get_group_id_get_local_size(int x, __global char *data)
 {
-    data[get_global_id(0)] = (ulong)get_group_id(0) + (-2);
-    data[get_global_id(1)] = (ulong)get_group_id(1) + (-16);
-    data[get_global_id(2)] = (ulong)get_group_id(2) + (-2);
+    data[get_global_id(0)] = (ulong)get_group_id(0) + (ulong)(-2);
+    data[get_global_id(1)] = (ulong)get_group_id(1) + (ulong)(-16);
+    data[get_global_id(2)] = (ulong)get_group_id(2) + (ulong)(-2);
 }
 
 __kernel __attribute__((reqd_work_group_size(2, 2, 4)))
@@ -1837,17 +1837,17 @@ void subtract_char_get_global_id_get_local_size(int x, __global uchar *data)
 __kernel __attribute__((reqd_work_group_size(16, 2, 2)))
 void subtract_char_get_global_size_get_local_size(__global char *data, int x)
 {
-    data[get_global_id(0)] = (ulong)get_global_size(0) + (-16);
-    data[get_global_id(1)] = (ulong)get_global_size(1) + (-2);
-    data[get_global_id(2)] = (ulong)get_global_size(2) + (-2);
+    data[get_global_id(0)] = (ulong)get_global_size(0) + (ulong)(-16);
+    data[get_global_id(1)] = (ulong)get_global_size(1) + (ulong)(-2);
+    data[get_global_id(2)] = (ulong)get_global_size(2) + (ulong)(-2);
 }
 
 __kernel __attribute__((reqd_work_group_size(2, 2, 4)))
 void subtract_char_get_num_groups_get_local_size(int x, __global uchar *data)
 {
-    data[get_global_id(0)] = (ulong)get_num_groups(0) + (-2);
-    data[get_global_id(1)] = (ulong)get_num_groups(1) + (-2);
-    data[get_global_id(2)] = (ulong)get_num_groups(2) + (-4);
+    data[get_global_id(0)] = (ulong)get_num_groups(0) + (ulong)(-2);
+    data[get_global_id(1)] = (ulong)get_num_groups(1) + (ulong)(-2);
+    data[get_global_id(2)] = (ulong)get_num_groups(2) + (ulong)(-4);
 }
 
 __kernel __attribute__((reqd_work_group_size(2, 2, 4)))
